Haram

+++ src/main/java/project/web/mvc/repository/AuthorityRepository.java
@@ -1,9 +1,12 @@
 package project.web.mvc.repository;
 
+import java.util.List;
+
 import org.springframework.data.repository.CrudRepository;
 
 import project.web.mvc.domain.Authority;
 
-public interface AuthorityRepository extends CrudRepository<Authority, Integer> {
+public interface AuthorityRepository extends CrudRepository<Authority, Long> {
 
+	
 }
